# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RR Plastic Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  store: "store",
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export interface DashboardSummary {
  totalMaterials: number;
  totalStockInToday: number;
  totalStockOutToday: number;
  lowStockCount: number;
  presentToday: number;
  absentToday: number;
  totalEmployees: number;
  totalCustomers: number;
  totalChallansToday: number;
}

export interface StockIn {
  id: number;
  materialId: number;
  materialName: string;
  quantity: number;
  date: string;
  supplier?: string;
  remarks?: string;
  createdAt: string;
}

export interface StockOut {
  id: number;
  materialId: number;
  materialName: string;
  quantity: number;
  date: string;
  purpose?: string;
  remarks?: string;
  createdAt: string;
}

export interface DeliveryChallanItem {
  id: number;
  materialId: number;
  materialName: string;
  description?: string;
  quantity: number;
  rate: number;
  amount: number;
}

export interface DeliveryChallan {
  id: number;
  challanNumber: string;
  customerId: number;
  customerName: string;
  vehicleNumber: string;
  date: string;
  totalAmount: number;
  items: DeliveryChallanItem[];
  createdAt: string;
}

export interface RecentActivity {
  stockIn: StockIn[];
  stockOut: StockOut[];
  challans: DeliveryChallan[];
}

export interface Material {
  id: number;
  name: string;
  unit: string;
  currentStock: number;
  lowStockThreshold: number;
  createdAt: string;
}

export interface CreateMaterialBody {
  name: string;
  unit: string;
  currentStock: number;
  lowStockThreshold: number;
}

export interface CreateStockInBody {
  materialId: number;
  quantity: number;
  date: string;
  supplier?: string;
  remarks?: string;
}

export interface CreateStockOutBody {
  materialId: number;
  quantity: number;
  date: string;
  purpose?: string;
  remarks?: string;
}

export interface Employee {
  id: number;
  name: string;
  role: string;
  phone: string;
  dailySalary: number;
  isActive: boolean;
  createdAt: string;
}

export interface CreateEmployeeBody {
  name: string;
  role: string;
  phone: string;
  dailySalary: number;
  isActive?: boolean;
}

export type AttendanceStatus =
  (typeof AttendanceStatus)[keyof typeof AttendanceStatus];

export const AttendanceStatus = {
  P: "P",
  A: "A",
  H: "H",
  L: "L",
} as const;

export interface Attendance {
  id: number;
  employeeId: number;
  employeeName: string;
  date: string;
  status: AttendanceStatus;
  createdAt: string;
}

export type CreateAttendanceBodyStatus =
  (typeof CreateAttendanceBodyStatus)[keyof typeof CreateAttendanceBodyStatus];

export const CreateAttendanceBodyStatus = {
  P: "P",
  A: "A",
  H: "H",
  L: "L",
} as const;

export interface CreateAttendanceBody {
  employeeId: number;
  date: string;
  status: CreateAttendanceBodyStatus;
}

export interface SalaryReportItem {
  employeeId: number;
  employeeName: string;
  role: string;
  dailySalary: number;
  totalDays: number;
  presentDays: number;
  halfDays: number;
  absentDays: number;
  leaveDays: number;
  finalSalary: number;
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  gst?: string;
  address?: string;
  createdAt: string;
}

export interface CreateCustomerBody {
  name: string;
  phone: string;
  gst?: string;
  address?: string;
}

export interface CustomerSummary {
  customerId: number;
  customerName: string;
  totalMaterialIssued: number;
  totalProductionReceived: number;
  balance: number;
}

export interface Production {
  id: number;
  customerId: number;
  customerName: string;
  materialId: number;
  materialName: string;
  quantityIssued: number;
  quantityReceived: number;
  date: string;
  remarks?: string;
  createdAt: string;
}

export interface CreateProductionBody {
  customerId: number;
  materialId: number;
  quantityIssued: number;
  quantityReceived: number;
  date: string;
  remarks?: string;
}

export interface CreateDeliveryChallanItemBody {
  materialId: number;
  description?: string;
  quantity: number;
  rate: number;
}

export interface CreateDeliveryChallanBody {
  customerId: number;
  vehicleNumber: string;
  date: string;
  items: CreateDeliveryChallanItemBody[];
}

export type MonthlyReportStockInByMaterialItem = {
  materialId: number;
  materialName: string;
  totalQty: number;
};

export type MonthlyReportStockOutByMaterialItem = {
  materialId: number;
  materialName: string;
  totalQty: number;
};

export interface MonthlyReport {
  month: number;
  year: number;
  totalStockIn: number;
  totalStockOut: number;
  materialBalance: number;
  totalSalaryPayable: number;
  stockInByMaterial: MonthlyReportStockInByMaterialItem[];
  stockOutByMaterial: MonthlyReportStockOutByMaterialItem[];
  customerSummary: CustomerSummary[];
  salaryReport: SalaryReportItem[];
}

export type ListMaterialsParams = {
  search?: string;
};

export type ListStockInParams = {
  startDate?: string;
  endDate?: string;
  materialId?: number;
};

export type ListStockOutParams = {
  startDate?: string;
  endDate?: string;
  materialId?: number;
};

export type ListEmployeesParams = {
  search?: string;
};

export type ListAttendanceParams = {
  employeeId?: number;
  month?: number;
  year?: number;
  date?: string;
};

export type GetSalaryReportParams = {
  month: number;
  year: number;
};

export type ListCustomersParams = {
  search?: string;
};

export type ListProductionParams = {
  customerId?: number;
  startDate?: string;
  endDate?: string;
};

export type ListDeliveryChallansParams = {
  customerId?: number;
  startDate?: string;
  endDate?: string;
};

export type GetMonthlyReportParams = {
  month: number;
  year: number;
};