# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RR Plastic Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List all materials
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create a material
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaterialBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [materials]
      summary: Get a material by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    put:
      operationId: updateMaterial
      tags: [materials]
      summary: Update a material
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaterialBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a material
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      security:
        - bearerAuth: []
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Create or update attendance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceBody"
      responses:
        "200":
          description: Attendance saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendance/{id}:
    delete:
      operationId: deleteAttendance
      tags: [attendance]
      summary: Delete attendance record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/summary:
    get:
      operationId: getCustomerSummary
      tags: [customers]
      summary: Get customer material/production summary
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSummary"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, store]
      required: [id, username, role]
...
    Material:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        lowStockThreshold:
          type: number
        createdAt:
          type: string
      required: [id, name, unit, currentStock, lowStockThreshold, createdAt]
    CreateMaterialBody:
      type: object
      properties:
        name:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        lowStockThreshold:
          type: number
      required: [name, unit, currentStock, lowStockThreshold]
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        dailySalary:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, role, phone, dailySalary, isActive, createdAt]
    CreateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        dailySalary:
          type: number
        isActive:
          type: boolean
      required: [name, role, phone, dailySalary]
    Attendance:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [P, A, H, L]
        createdAt:
          type: string
      required: [id, employeeId, employeeName, date, status, createdAt]
    CreateAttendanceBody:
      type: object
      properties:
        employeeId:
          type: integer
        date:
          type: string
        status:
          type: string
          enum: [P, A, H, L]
      required: [employeeId, date, status]
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        gst:
          type: string
        address:
          type: string
        createdAt:
          type: string
      required: [id, name, phone, createdAt]
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        gst:
          type: string
        address:
          type: string
      required: [name, phone]
    CustomerSummary:
      type: object
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        totalMaterialIssued:
          type: number
        totalProductionReceived:
          type: number
        balance:
          type: number
      required: [customerId, customerName, totalMaterialIssued, totalProductionReceived, balance]
...